# src/swaggerdoc.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Todo Backend",
        "description": "Todos APIs",
        "contact": {
            "email": "fidelis.ojeah@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/v1"
        }
    ],
    "paths": {
        "/users": {
            "post": {
                "tags": ["users"],
                "summary": "Create a new user",
                "description": "create a new user, email is unique",
                "operationId": "createUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignupRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserAPIResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid types supplied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponseError400"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": ["users"],
                "summary": "Modify a user",
                "description": "modify an existing user",
                "operationId": "updateUser",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserModificationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserAPIResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid types supplied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponseError400"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unable to authorize user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponseError401"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["users"],
                "summary": "Modify a user",
                "description": "modify an existing user",
                "operationId": "updateUser",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserModificationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserAPIResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid types supplied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponseError400"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unable to authorize user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponseError401"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/auth": {
            "post": {
                "tags": ["user login"],
                "summary": "Logs a user in",
                "description": "Logs a user into system, response is token",
                "operationId": "loginUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginAPIResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid types supplied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponseError400"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "APIResponseError401": {
                "type": "object",
                "properties": {
                    "VERSION": {
                        "type": "string"
                    },
                    "statusCode": {
                        "type": "number",
                        "format": "int64"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {}
                        }
                    }
                }
            },
            "APIResponseError400": {
                "type": "object",
                "properties": {
                    "VERSION": {
                        "type": "string"
                    },
                    "statusCode": {
                        "type": "number",
                        "format": "int64"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {}
                        }
                    }
                }
            },
...
            "UserAPIResponse": {
                "type": "object",
                "properties": {
                    "VERSION": {
                        "type": "string"
                    },
                    "statusCode": {
                        "type": "number",
                        "format": "int64"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserResponse"
                        }
                    }
                }
            },
            "LoginAPIResponse": {
                "type": "object",
                "properties": {
                    "VERSION": {
                        "type": "string"
                    },
                    "statusCode": {
                        "type": "number",
                        "format": "int64"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LoginResponse"
                        }
                    }
                }
            },
...
            "SignupRequest": {
                "type": "object",
                "properties": {
                    "profile": {
                        "type": "object",
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "gender": {
                                "type": "string"
                            },
                            "timezone": {
                                "type": "string"
                            }
                        }
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "UserResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "timezone": {
                        "type": "string"
                    }
                }
            },
            "UserLoginRequest": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "LoginResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "timezone": {
                        "type": "string"
                    },
                    "token": {
                        "type": "string"
                    }
                }
            },
            "UserModificationRequest": {
                "type": "object",
                "properties": {
                    "profile": {
                        "type": "object",
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "gender": {
                                "type": "string"
                            },
                            "timezone": {
                                "type": "string"
                            }
                        }
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    }
                }
            }
        }
    },
...
}